fix(deps): migrate docs site to Astro v7 (CVE-2026-33769)
Resolves the security update flagged by Renovate (astro 5.17 -> 7,
GHSA-g735-7g2w-hh3f / CVE-2026-33769 matchPathname).

- astro 7.3.2 + @astrojs/react 6 + @astrojs/sitemap 3.7.4
- replace deprecated @astrojs/tailwind with @tailwindcss/vite and migrate
  global.css to Tailwind v4 CSS-first config (@custom-variant dark +
  @theme inline); HSL token values unchanged so visuals are identical
- migrate legacy content collections (src/content/config.ts) to the
  Content Layer glob() loader in src/content.config.ts (legacy API
  removed in Astro v6); entry ids stay the yaml file stems
- drop tailwindcss-animate (no animate-* utilities used in docs)
- make the BaseLayout path prefix unconditional: Astro v6+ serves
  `astro dev` under the base path, so the old isProd ternary 404'd the
  language switcher in dev
- docs workflow: Node 20 -> 24 (Astro v7 requires >= 22.12)

Closes #642
