+++ docs/src/components/animate-ui/components/radix/hover-card.tsx
@@ -33,7 +33,7 @@ export function HoverCardContent({
         align={align}
         sideOffset={sideOffset}
         className={cn(
-          "origin-(--radix-hover-card-content-transform-origin) outline-hidden z-50 w-64 rounded-md border bg-popover p-4 text-popover-foreground shadow-md",
+          "bg-popover text-popover-foreground z-50 w-64 origin-(--radix-hover-card-content-transform-origin) rounded-md border p-4 shadow-md outline-hidden",
           className,
         )}
         {...props}
