+++ docs/src/components/layouts/Footer.astro
@@ -8,7 +8,7 @@ const currentYear = new Date().getFullYear();
 ---
 
 <footer class="border-t py-6">
-  <div class="container mx-auto px-4 text-center text-sm text-muted-foreground">
+  <div class="text-muted-foreground container mx-auto px-4 text-center text-sm">
     <p>© {currentYear} j4rviscmd</p>
   </div>
 </footer>
