+++ docs/src/components/sections/Hero.astro
@@ -17,7 +17,7 @@ const t = useTranslations(lang);
   <h1 class="text-4xl font-bold tracking-tight sm:text-5xl md:text-6xl">
     {t("hero.title")}
   </h1>
-  <p class="mx-auto mt-6 max-w-2xl text-lg text-muted-foreground">
+  <p class="text-muted-foreground mx-auto mt-6 max-w-2xl text-lg">
     {t("hero.description")}
   </p>
   <div class="mt-10 flex w-full flex-wrap justify-center gap-4">
